docs/submitting: remove reference to old Zenodo version
Upon acceptance, authors archive their repositories to Zenodo, but
that will be a new version not available when writing the paper.  In
the provided example, the @zenodo reference in the paper is for v0.2.1
while the archive in the sidebar is v0.2.2 (the post-review version).
The citation to the old version does not add value, but may cause
confusion.

Reported-by: Gregory R. Lee <grlee77@gmail.com>